pyproject.toml
[project]
name = "gde-ai"
version = "0.1.0"
description = "Example on Pydantic AI"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.12.4",
    "pydantic-ai>=1.11.1",
    "python-dotenv>=1.2.1",
]
